Remove `shader.rs` from `wgpu`
