+++ compositor_render/src/transformations/shader/pipeline.rs
@@ -6,8 +6,7 @@ use crate::{
     scene::ShaderParam,
     state::render_graph::NodeId,
     wgpu::{
-        common_pipeline::{Sampler, Vertex},
-        shader::{CreateShaderError, FRAGMENT_ENTRYPOINT_NAME, VERTEX_ENTRYPOINT_NAME},
+        common_pipeline::{CreateShaderError, Sampler, Vertex},
         texture::{NodeTexture, NodeTextureState, RGBATexture},
         validation::{validate_contains_header, ParametersValidationError},
         WgpuCtx, WgpuErrorScope,
@@ -91,11 +90,11 @@ impl ShaderPipeline {
                 vertex: wgpu::VertexState {
                     buffers: &[Vertex::LAYOUT],
                     module: &shader_module,
-                    entry_point: VERTEX_ENTRYPOINT_NAME,
+                    entry_point: crate::wgpu::common_pipeline::VERTEX_ENTRYPOINT_NAME,
                 },
                 fragment: Some(wgpu::FragmentState {
                     module: &shader_module,
-                    entry_point: FRAGMENT_ENTRYPOINT_NAME,
+                    entry_point: crate::wgpu::common_pipeline::FRAGMENT_ENTRYPOINT_NAME,
                     targets: &[Some(wgpu::ColorTargetState {
                         format: wgpu::TextureFormat::Rgba8Unorm,
                         write_mask: wgpu::ColorWrites::all(),
