+++ compositor_render/src/transformations/web_renderer/shader.rs
@@ -3,8 +3,7 @@ use std::sync::Arc;
 use wgpu::ShaderStages;
 
 use crate::wgpu::{
-    common_pipeline::{Sampler, Vertex},
-    shader::{CreateShaderError, FRAGMENT_ENTRYPOINT_NAME, VERTEX_ENTRYPOINT_NAME},
+    common_pipeline::{CreateShaderError, Sampler, Vertex},
     texture::{NodeTextureState, Texture},
     WgpuCtx, WgpuErrorScope,
 };
@@ -72,11 +71,11 @@ impl WebRendererShader {
                 vertex: wgpu::VertexState {
                     buffers: &[Vertex::LAYOUT],
                     module: &shader_module,
-                    entry_point: VERTEX_ENTRYPOINT_NAME,
+                    entry_point: crate::wgpu::common_pipeline::VERTEX_ENTRYPOINT_NAME,
                 },
                 fragment: Some(wgpu::FragmentState {
                     module: &shader_module,
-                    entry_point: FRAGMENT_ENTRYPOINT_NAME,
+                    entry_point: crate::wgpu::common_pipeline::FRAGMENT_ENTRYPOINT_NAME,
                     targets: &[Some(wgpu::ColorTargetState {
                         format: wgpu::TextureFormat::Rgba8Unorm,
                         write_mask: wgpu::ColorWrites::all(),
