+++ compositor_render/src/wgpu.rs
@@ -3,7 +3,6 @@ use log::error;
 pub(crate) mod common_pipeline;
 mod ctx;
 pub(crate) mod format;
-pub(crate) mod shader;
 pub(crate) mod texture;
 pub(crate) mod utils;
 pub(crate) mod validation;
