+++ compositor_render/src/wgpu/validation/error.rs
@@ -1,4 +1,6 @@
-use super::{USER_DEFINED_BUFFER_BINDING, USER_DEFINED_BUFFER_GROUP, VERTEX_ENTRYPOINT_NAME};
+use crate::wgpu::common_pipeline::{
+    USER_DEFINED_BUFFER_BINDING, USER_DEFINED_BUFFER_GROUP, VERTEX_ENTRYPOINT_NAME,
+};
 
 const HEADER_DOCS_URL: &str =
     "https://github.com/membraneframework/video_compositor/wiki/Shader#header";
