+++ compositor_render/src/wgpu/validation/shader_header.wgsl
